日志(修复): hardware_concurrency 按 size_t 转换——%zu 格式不再读越界
std::thread::hardware_concurrency() 返回 unsigned(32位)，%zu 按 64 位
size_t 打印导致高位为寄存器残留垃圾（实测显示 2405181685792，
真实值 32 藏在低位）。纯诊断日志显示问题，不影响功能。

+++ face_service/onnx_models.cpp
@@ -103,7 +103,8 @@ bool OnnxRecognizer::Initialize(const std::wstring& modelPath) {
         // on E-cores. Knowing the actual value distinguishes "few threads"
         // from "threads present but slow" in one log line.
         FACELOGIN_INFO(L"  ONNX intra-op threads: %d (hardware_concurrency=%zu)",
-                       OnnxThreadCount(), std::thread::hardware_concurrency());
+                       OnnxThreadCount(),
+                       static_cast<size_t>(std::thread::hardware_concurrency()));
 
         m_initialized = true;
         return true;
